engine: extract esm/settings.js
